+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to `homebridge` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## v2.2.1 (Pending Release)
+
+### Changes
+
+- fix(matter): add electrical clusters to public cluster name map
+
+### Homebridge Dependencies
+
+- `@homebridge/hap-nodejs` @ `v2.1.9`
+
+### Matter Dependencies
+
+- `@matter/main` @ `v0.17.6`
+
 ## v2.2.0 (2026-07-18)
 
 ### Changes